yesterday's work

+++ HostelWorld/src/main/java/edu/nju/hostelworld/service/UserServiceImpl.java
@@ -17,19 +17,48 @@ public class UserServiceImpl implements UserService {
     private UserDao userDao;
 
     public User saveUser(User user) {
-        return userDao.save(user);
+        if(!isUsernameExist(user.getUsername())&&!isCardExist(user.getCardId())){
+            return userDao.save(user);
+        }
+        return null;
     }
+
 
     public void deleteUser(String username) {
         User user = userDao.findByUsername(username);
-        userDao.delete(user);
+        if(user != null) {
+            userDao.delete(user);
+        }
     }
 
     public User findUser(String username, String password) {
         return userDao.findByUsernameAndPassword(username,password);
     }
 
+    public User findUser(String username) {
+        return userDao.findByUsername(username);
+    }
+
     public List<User> findAllUser() {
         return userDao.findAll();
+    }
+
+
+    private boolean isUsernameExist(String username){
+        if(userDao.findByUsername(username) == null){
+            return false;
+        }else{
+            return true;
+        }
+
+    }
+
+    public boolean isCardExist(String cardId){
+        if(userDao.findByCardId(cardId) == null){
+            return false;
+        }else{
+            return true;
+        }
+
     }
 }
